% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{adaptative_threshold_otu}
\alias{adaptative_threshold_otu}
\title{Relevel taxa/OTU to 0 according to a taxa/OTU wise threshold.}
\usage{
adaptative_threshold_otu(physeq, otuThreshold)
}
\arguments{
\item{physeq}{\code{\link[phyloseq]{otu_table-class}}, or \code{\link[phyloseq]{phyloseq-class}}}

\item{otuThreshold}{A \code{\link{numeric-class} of length equal to the number of taxa/OTU.
If the latter is zero, the corresponding taxa/OTU will not be relevel.}}
}
\value{
An object of class \code{\link[phyloseq]{phyloseq}}.
 Only the \code{\link[phyloseq]{otu_table-class}} component is modified.
}
\description{
Relevel taxa/OTU to 0 according to a taxa/OTU wise threshold.
}
\references{
This procedure is described in \href{https://doi.org/10.1128/mSystems.00032-16}{Galan et al. (2016)}.
}
