% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_stats.R
\name{ctrl_samples_stats}
\alias{ctrl_samples_stats}
\title{Generate taxa/OTU statistics to compare control samples with the remaining dataset}
\usage{
ctrl_samples_stats(physeq, physeq_ctrl)
}
\arguments{
\item{physeq}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}}

\item{physeq_ctrl}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}
of control sample. It can be the output of \code{\link{subset_samples_no_zero}}
 for example.}
}
\value{
A \code{\link{data.frame}} with the following components
\describe{
\item{\code{RA_In_Ctrl}}{Taxa/OTU relative abundance in Control sample}
\item{\code{RA_In_All}}{Same Taxa/OTU relative abundance in all samples \strong{but} Control samples}
\item{\code{Ratio}}{Ratio of \code{RA_In_Ctrl}/\code{RA_In_All}}
\item{\code{Prevalence}}{Taxa/OTU Prevalence in all samples \strong{but} Control samples.
 Range from 0 to the number of samples in \code{physeq} object
  (obtained by \code{\link[phyloseq]{ntaxa}})}
}
}
\description{
Generate taxa/OTU statistics to compare control samples with the remaining dataset
}
\details{
The taxa/OTU prevalence is defined as the number of samples in which it occurs.
The prevalence is corrected by substracting the number of samples in the \code{physeq_ctrl}
object (usually 1).
}
\examples{
data(soilrep)
# Subset a sample
soilrep.6CC<-subset_samples_no_zero(soilrep, Sample=="6CC")
# Check
ctrl_df<-ctrl_samples_stats(soilrep,soilrep.6CC)
head(ctrl_df)
}

