% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{get_taxa_nomy}
\alias{get_taxa_nomy}
\title{Taxa/OTU abundance in a given sample coupled with rank and taxonomic information}
\usage{
get_taxa_nomy(physeq, smp, level = "Species")
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}}}

\item{smp}{A single sample of interest. This integer or sample name will be
passed to \code{\link[phyloseq]{get_taxa}}.}

\item{level}{A character indicating the taxonomic rank of interest. Should be in
\code{\link[phyloseq]{rank_names}}. Default is "Species".}
}
\value{
An abundance sorted \code{\link{data.frame}} with the following components
\describe{
\item{\code{Sample}}{Sample of interest}
\item{\code{Abundance}}{Taxa/OTU abundance in the sample of interest}
\item{\code{Rank}}{Taxa/OTU rank in the dataset. The higher the rank, the less abundant the OTU.}
\item{\code{level}}{Taxonomic rank of interest. Note that the column is named after
the provided value from \code{level}.}
}
}
\description{
Taxa/OTU abundance in a given sample coupled with rank and taxonomic information
}
\seealso{
\code{\link[phyloseq]{get_taxa}}
}
