% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{id2hash}
\alias{id2hash}
\title{Convert taxa/ID through a correspondence table}
\usage{
id2hash(physeq, hashTable, ...)
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}}}

\item{hashTable}{A \code{\link{data.frame}} with at least a **UsearchID** and **HashID** columns.}

\item{...}{The subsetting expression that should be applied to the
correspondance table \code{hashTable}. This is passed on to \code{\link[base]{subset}}.}
}
\value{
physeq-renamed \code{\link{phyloseq-class}} with OTU renamed.
}
\description{
Convert taxa/ID through a correspondence table
}
