% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{multiple_rrarefy}
\alias{multiple_rrarefy}
\title{Performs rarefying multiple times and returns averaged OTU table}
\usage{
multiple_rrarefy(physeq, sample.size = min(sample_sums(physeq)), n,
  start.seed)
}
\arguments{
\item{physeq}{(Required). A \code{\link{phyloseq-class}} object that you
want to trim/filter.}

\item{sample.size}{(Optional). A single integer value equal to the number
of reads being simulated, also known as the depth,
and also equal to each value returned by \code{\link{sample_sums}}
on the output.}

\item{n}{The number of times to perform rarefying of OTU table}

\item{start.seed}{Random seed to start, it will be incremented up to \code{start.seed+n}}
}
\value{
An object of class \code{\link[phyloseq]{phyloseq}}.
 Only the \code{\link[phyloseq]{otu_table-class}} component is modified.
}
\description{
Performs rarefying multiple times and returns averaged OTU table
}
\seealso{
\code{\link[phyloseq]{rarefy_even_depth}}
}
