% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_krona}
\alias{plot_krona}
\title{Interactive Taxonomy plot with Krona from a phyloseq object}
\usage{
plot_krona(physeq, output)
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}} with a \code{\link{taxonomyTable-class}}}

\item{output}{A \code{\link{character}} stating the output filename for Krona Chart.}
}
\description{
Construct and run a Krona Chart to compare taxonomic assignations between
the different conditions described in the \code{Description} variable (for now).
}
\examples{
\dontrun{
require(phyloseq)
data(GlobalPatterns)
# Little hack before being able to generalize plot_krona
sample_data(GlobalPatterns)$Description<-sample_data(GlobalPatterns)$SampleType
plot_krona(GlobalPatterns,"GP-krona")
}
}
\references{
\href{https://github.com/marbl/Krona/wiki/KronaTools}{KronaTools}.
}
