% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mv}
\alias{plot_mv}
\title{Plot Mean Variance OTU in order to evaluate overdispersion}
\usage{
plot_mv(physeq, log = TRUE, title = NULL)
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}}}

\item{log}{A \code{\link{logical}} stating whether to apply \code{\link{log1p}} transform to both axis.}

\item{title}{A title to the mean-variance plot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
The \code{\link{mean}} and variance (\code{\link{var}}) of \code{physeq} OTU table are computed.
Overdispersion can be visually assessed by the generated \code{\link{plot}}. The putative linear relation
between mean and variance is added with \code{\link{abline}} to guide the analysis.
}
\details{
The \code{\link{log1p}} transform was chosen to limit \code{\link{NaN}} values with null values.
}
\examples{
require(phyloseq)
data(GlobalPatterns)
(mv1<-plot_mv(GlobalPatterns, title = "Mean-Variance Plot with log-transform"))
mv1.nolog<-plot_mv(GlobalPatterns,log = FALSE, title = "Mean-Variance Plot without log-transform")
}
\references{
Modified from \href{https://doi.org/10.1186/s40168-016-0208-8}{Thorsen and Brejnrod et al. (2016)}.
}
\seealso{
\code{\link{taxa_stats}}
}

