% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_pie}
\alias{plot_pie}
\title{Pie Plot Taxonomy Assignations from Phyloseq}
\usage{
plot_pie(physeq, title., condition, COL.TAXO. = COL.TAXO)
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}} with a \code{\link{sample_data-class}}}

\item{title.}{A \code{\link{character}} specifying the plot title}

\item{condition}{A \code{\link{character}} value of the "Description" column in \code{\link{sample_data-class}}}

\item{COL.TAXO.}{COL.TAXO A palette (\code{\link{list}}) containing (at least) two named elements}
}
\value{
A \code{\link{list}} of two elements (\code{inner} and \code{outer}) containing taxa names.
}
\description{
Displays an overlay of two pie chart: (i) inner (e.g. \href{https://en.wikipedia.org/wiki/Phylum}{Phylum})
and (ii) outer (e.g. \href{https://en.wikipedia.org/wiki/Class_(biology)}{Class})
}
\details{
It was primarily developed to illustrate taxonomic assignations at both
levels at the same time.
}
\references{
Adapted from Fort et al. (2016) in \href{https://doi.org/10.7717/peerj.2656}{PeerJ}.
}
\seealso{
\code{\link{plot_pie_vector}}, \code{\link{psmelt}}
}
\author{
Charlie PAUVERT
}
