% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_pie_vector}
\alias{plot_pie_vector}
\title{Pie Plot Taxonomy Assignations from Two Abundance Vector}
\usage{
plot_pie_vector(inner_abd, outer_abd, title, COL.TAXO = COL.TAXO, EDG = 200)
}
\arguments{
\item{inner_abd}{A \code{\link{numeric-class}} describing the inner pie chart}

\item{outer_abd}{A \code{\link{numeric-class}} describing the outer pie chart}

\item{title}{A \code{\link{character}} specifying the plot title}

\item{COL.TAXO}{A palette (\code{\link{list}}) containing (at least) two named elements}

\item{EDG}{Number of edges}
}
\value{
Nothing
}
\description{
Displays an overlay of two pie chart: (i) inner and (ii) outer.
It was primarily developed to illustrate taxonomic assignations at both
levels at the same time
(e.g. \href{https://en.wikipedia.org/wiki/Phylum}{Phylum} and
 \href{https://en.wikipedia.org/wiki/Class_(biology)}{Class})
}
\details{
\code{COL.TAXO} should contain at least two named elements \code{DIV} and \code{CLASS}.
These vectors describe the inner and outer pie chart color, respectively.
}
\references{
See Fort et al. (2016) in \href{https://doi.org/10.7717/peerj.2656}{PeerJ}.
}
\seealso{
\code{\link{floating.pie}} (in \code{\link{plotrix}} package)
}
\author{
Thomas Fort
}
