% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_pie_with_legend}
\alias{plot_pie_with_legend}
\title{Plot Taxonomy Assignations from Phyloseq with legend.}
\usage{
plot_pie_with_legend(physeq, title., condition)
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}} with a \code{\link{sample_data-class}}}

\item{title.}{A \code{\link{character}} specifying the plot title}

\item{condition}{A \code{\link{character}} value of the "Description" column in \code{\link{sample_data-class}}}
}
\description{
Plot Taxonomy Assignations from Phyloseq with legend.
}
\seealso{
\code{\link{plot_pie}}
}

