% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_sparsity}
\alias{plot_sparsity}
\title{Plot taxa/OTU table to evaluate sparsity.}
\usage{
plot_sparsity(physeq, title = NULL)
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}}}

\item{title}{A title to the mean-variance plot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Taxa/OTU are ordered by their prevalence in the dataset, and samples are
ordered by their richness (or completedness).
}
\examples{
\dontrun{
require(phyloseq)
data(enterotype)
(entero.sparse<-plot_sparsity(enterotype, title = "Sparsity Plot")
}
}
\references{
Adapted from \href{https://doi.org/10.1186/s40168-016-0208-8}{Thorsen and Brejnrod et al. (2016)}.
}
\seealso{
\code{\link{plot_mv}},\code{\link{taxa_prev}}, \code{\link{sample_prev}}
}

