% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_violin}
\alias{plot_violin}
\title{Violin distribution plot}
\usage{
plot_violin(physeq, x, title = NULL)
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}}}

\item{x}{A \code{\link{character}} describing the variable from
the \code{\link[phyloseq]{sample_data}} object to plot against.}

\item{title}{A title to the mean-variance plot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Violin plot to visualise sample sequence number distribution.
}
\examples{
require(phyloseq)
data("soilrep")
head( sample_data(soilrep) )
plot_violin(physeq = soilrep, x = "warmed")
}

