% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_stats.R
\name{relative_abd_without_specific_sample}
\alias{relative_abd_without_specific_sample}
\title{Compute taxa/OTU abundance relative to the total dataset}
\usage{
relative_abd_without_specific_sample(physeq, specific)
}
\arguments{
\item{physeq}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}}

\item{specific}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}
containing samples of interest to examine (e.g. Negative or Positive Controls).}
}
\value{
A \code{\link{numeric-class}} with length equal to the number of species
 present in the second --\code{specific}-- \code{phyloseq} object.
  Names indicate the taxa ID, and value equal to the taxa/OTU
 relative abundance in the first --\code{physeq}-- object (pruned of specific samples).
}
\description{
Convenience function to divide \code{taxa_sums} output vector by the
total number of sequences.
Samples present in the second \code{phyloseq} object --\code{specific}-- are discarded
 from the first one --\code{physeq}.
Taxa/OTU present in the second \code{phyloseq} object --\code{specific}-- and supposedly
of interest are stored.
The relative abundances of these taxa/OTU of interest are then computed.
}
\seealso{
\code{\link{taxa_sums}} on which this function is based.
See \code{\link{ctrl_samples_stats}} for a usage example.
}
