% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bak_stats.r, R/simple_stats.R
\name{relative_abundance}
\alias{relative_abundance}
\title{Compute taxa/OTU relative abundance}
\usage{
relative_abundance(physeq)

relative_abundance(physeq)
}
\arguments{
\item{physeq}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}}

\item{physeq}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}}
}
\value{
A \code{\link{numeric-class}} with length equal to the number of species
 in the table, name indicated the taxa ID, and value equal to the taxa/OTU
 relative abundance

A \code{\link{numeric-class}} with length equal to the number of species
 in the table, name indicated the taxa ID, and value equal to the taxa/OTU
 relative abundance
}
\description{
Compute taxa/OTU relative abundance

Convenience function to divide \code{taxa_sums} output vector by the
total number of sequences.
}
\seealso{
\code{\link{taxa_sums}} on which this function is based.
See \code{\link{ctrl_samples_stats}} for a usage example.
}

