% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_stats.R
\name{sample_prev}
\alias{sample_prev}
\title{Taxa/OTU number in all samples}
\usage{
sample_prev(physeq)
}
\arguments{
\item{physeq}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}}
}
\value{
A \code{\link{numeric-class}} with length equal to the number of samples
 in the table, name indicated the sample ID, and value equal to the number of taxa/OTU
 each samples contains.
}
\description{
Return a named vector of samples with the number of taxa/OTU where
their abundance was not null. It is similar to the observed richness.
}
\examples{
require(phyloseq)
data(esophagus)
sample_prev(esophagus)
}
\seealso{
\code{\link{taxa_sums}} on which this function is based.
 \code{\link{taxa_prev}} for taxa/OTU prevalence across samples.
}
