% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_stats.R
\name{sample_stats}
\alias{sample_stats}
\title{Apply statistics function to taxa abundance vector}
\usage{
sample_stats(physeq, ...)
}
\arguments{
\item{physeq}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}}

\item{...}{The function to be applied to the OTU table}
}
\value{
A \code{\link{numeric-class}} with length equal to the number of samples
 in the table. Names indicate sample ID, and values equal to the result of
 the function applied through the passed expression (\code{...}).
}
\description{
It generalize the \code{\link[phyloseq]{sample_sums}} function and enable the use of several summary functions.
}
\examples{
require(phyloseq)
data(esophagus)
sample_stats(esophagus, mean)[1:6]
sample_stats(esophagus, var)[1:6]
}
\seealso{
\code{\link{sample_sums}} on which this function is based.
}

