% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_stats.R
\name{subset_samples_no_zero}
\alias{subset_samples_no_zero}
\title{Subset samples and prune taxa/OTU with zero abundance}
\usage{
subset_samples_no_zero(physeq, ...)
}
\arguments{
\item{physeq}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}}

\item{...}{The subsetting expression that should be applied to the
\code{sampleTable}. This is passed on to \code{\link{subset_samples}}.
Note that taxa/OTU with zero abundance are pruned from the \code{\link{phyloseq-class}} object.}
}
\value{
A subsetted object with the same class as \code{physeq}
}
\description{
Subset samples and prune taxa/OTU with zero abundance
}
\examples{
data(GlobalPatterns)
gp<-subset_samples(GlobalPatterns, SampleType=="Ocean")
gp.nz<-subset_samples_no_zero(GlobalPatterns, SampleType=="Ocean")
ntaxa(gp) # 19216 OTU
ntaxa(gp.nz) # 5669 OTU
}
\seealso{
\code{\link{subset_samples}} on which it rely.
}

