% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_stats.R
\name{taxa_prev}
\alias{taxa_prev}
\title{OTU/Species prevalence in all samples}
\usage{
taxa_prev(physeq)
}
\arguments{
\item{physeq}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}}
}
\value{
A \code{\link{numeric-class}} with length equal to the number of species
 in the table, name indicated the taxa ID, and value equal to the species prevalence
 in every samples where it occurs.
}
\description{
Return a named vector of taxa/OTU with the number of samples where
their abundance was not null.
}
\examples{
require(phyloseq)
data(esophagus)
taxa_prev(esophagus)
}
\seealso{
\code{\link{taxa_sums}} on which this function is based.
 \code{\link{sample_prev}} for a similar output on samples.
}

