% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_stats.R
\name{taxa_stats}
\alias{taxa_stats}
\title{Apply statistics function to taxa abundance vector}
\usage{
taxa_stats(physeq, ...)
}
\arguments{
\item{physeq}{\code{\link{otu_table-class}}, or \code{\link{phyloseq-class}}}

\item{...}{The function to be applied to the OTU table}
}
\value{
A \code{\link{numeric-class}} with length equal to the number of species
 in the table, name indicated the taxa ID, and value equal to the result of
 the function applied through \code{...}.
}
\description{
It generalize the \code{\link{taxa_sums}} function and enable the use of several summary functions.
}
\examples{
require(phyloseq)
data(esophagus)
taxa_stats(esophagus, mean)[1:6]
taxa_stats(esophagus, var)[1:6]
}
\seealso{
\code{\link{taxa_sums}} on which this function is based.
}

